requirements.txt
python-dotenv==0.21.1
pyotp==2.7.0
djongo==1.3.6
djangorestframework-simplejwt==5.2.0
django-dotenv==1.4.2
django-rest-framework-mongoengine==3.4.1
Django==3.2
django-cors-headers==3.13.0
PyJWT==2.4.0
gunicorn==20.1.0
psycopg2==2.9.5
pymongo==3.12.3
djangorestframework==3.14.0
dnspython==2.2.1
sqlparse==0.2.4